+++ boreas/cli.h
@@ -17,33 +17,13 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef BOREAS_IO_H
-#define BOREAS_IO_H
+#ifndef CLI_H
+#define CLI_H
 
-#include "../base/hosts.h"
-#include "../util/kb.h"
 #include "alivedetection.h"
 #include "boreas_error.h"
 
-gvm_host_t *
-get_host_from_queue (kb_t, gboolean *);
-
-void
-put_finish_signal_on_queue (void *);
-
-void
-send_dead_hosts_to_ospd_openvas (int);
-
-void
-init_scan_restrictions (struct scanner *, int);
-
-void
-handle_scan_restrictions (struct scanner *, gchar *);
-
-gchar *
-get_openvas_scan_id (const gchar *, int);
-
 boreas_error_t
-get_alive_test_methods (alive_test_t *);
+run_cli (gvm_hosts_t *, alive_test_t, const gchar *);
 
-#endif /* not BOREAS_IO_H */
+#endif /* not CLI_H */